Change index type to int from uint64_t.

+++ src/messages.h
@@ -16,8 +16,8 @@
 
 #include <libsoup/soup.h>
 
-void messages_add(SoupMessage *msg, uint64_t index);
+void messages_add(SoupMessage *msg, int index);
 
-SoupMessage *message_get(uint64_t index);
+SoupMessage *message_get(int index);
 
 #endif
